Cody: Add jaccard-based context to inline completions
